requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
names==0.3.0
pokebase==1.3.0
py-randomizer==0.0.3
requests==2.28.1
urllib3==1.26.12